Resolve pr #45

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  after_action :verify_authorized
   before_action :authenticate_user_using_x_auth_token
   before_action :load_quiz, only: %i[create update]
   before_action :load_question, only: %i[destroy show update]
 
   def create
     question = Question.new(question_params)
+    authorize question.quiz
     if question.save!
       render status: :ok, json: { notice: t("successfully_created", entity: "Question") }
     else
@@ -16,10 +18,12 @@ class QuestionsController < ApplicationController
   end
 
   def show
+    authorize @question.quiz
     render status: :ok, json: { question: { question: @question.question, options: @question.options } }
   end
 
   def update
+    authorize @question.quiz
     if @question.update(question_params)
       render status: :ok, json: { notice: t("successfully_updated", entity: "Question") }
     else
@@ -29,6 +33,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
+    authorize @question.quiz
     if @question.destroy
       render status: :ok, json: { notice: t("successfully_deleted", entity: "Question") }
     else
